优化

+++ server/src/controller/_init.js
@@ -12,13 +12,20 @@ router.post('/init', async (ctx, next) => {
     if (!await mongodb.collection(Util.CollectionEnum.agent).findOne({ userName: 'admin' })) {
         // 创建集合
         for (let key in Util.CollectionEnum) {
-            await mongodb.dropCollection(Util.CollectionEnum[key])
-            await mongodb.createCollection(Util.CollectionEnum[key])
+            if (!key.startsWith('v')) {
+                await mongodb.dropCollection(Util.CollectionEnum[key])
+                await mongodb.createCollection(Util.CollectionEnum[key])
+            }
         }
         // 创建索引
         await createIndex(mongodb)
         // 预置数据
-        await mongodb.collection(Util.CollectionEnum._seq).insertMany([{ seqName: 'billSeq', seqValue: 0 }, { seqName: 'reviewSeq', seqValue: 0 }, { seqName: 'messageSeq', seqValue: 0 },{ seqName: 'profitSeq', seqValue: 0 }])
+        await mongodb.collection(Util.CollectionEnum._seq).insertMany([
+            { seqName: 'billSeq', seqValue: 0 },
+            { seqName: 'reviewSeq', seqValue: 0 },
+            { seqName: 'messageSeq', seqValue: 0 },
+            { seqName: 'profitSeq', seqValue: 0 }
+        ])
 
         await mongodb.collection(Util.CollectionEnum.agent).insertOne({ id: 100000, role: 'admin', userName: 'admin', userPwd: '123456', userNick: '超级管理员', subrole: '超级管理员', status: 1, createAt, createAtStr: moment(createAt).utcOffset(8).format('YYYY-MM-DD HH:mm:ss') })
         await mongodb.collection(Util.CollectionEnum.subrole).insertOne({ id: 1000, roleName: '超级管理员', permissions: ['所有权限', '代理中心', '玩家中心', '代理账单', '玩家账单', '审核中心', '管理中心', '管理员列表', '角色列表'], createAt, createAtStr: moment(createAt).utcOffset(8).format('YYYY-MM-DD HH:mm:ss') })
@@ -63,33 +70,34 @@ async function createIndex(mongodb) {
     await mongodb.collection(Util.CollectionEnum.bill).createIndex({ sourceRelKey: 1, project: 1 }, { sparse: true })
 
     await mongodb.collection(Util.CollectionEnum.review).createIndex({ id: -1 })
+    await mongodb.collection(Util.CollectionEnum.profit).createIndex({ id: -1 })
     await mongodb.collection(Util.CollectionEnum.message).createIndex({ id: -1 })
 }
 
-async function testTransaction() {
-    const mongodb = global.mongodb
-    const session = await global.getMongoSession()
-    try {
-        await mongodb.collection(Util.CollectionEnum._seq).findOneAndUpdate(
-            { seqName: 'billSeq' },
-            { $inc: { seqValue: 1 } },
-            { returnOriginal: false, projection: { seqValue: 1, _id: 0 }, session }
-        )
-        await mongodb.collection(Util.CollectionEnum.bill).insertOne({ id: 1, sourceId: 1 }, { session })
-        await mongodb.collection(Util.CollectionEnum.bill).insertOne({ id: 2 }, { session })
-        await mongodb.collection(Util.CollectionEnum.bill).insertOne({ id: 3 }, { session })
-        await mongodb.collection(Util.CollectionEnum.bill).insertOne({ id: 4, sourceId: 1 }, { session })
-        await session.commitTransaction()
-    } catch (error) {
-        console.error('事务回滚')
-        console.error(error.codeName)
-        console.error(error.errmsg) // sourceId_1
-        console.error('错误详情')
-        console.error(error)
-        await session.abortTransaction()
-    } finally {
-        await session.endSession()
-    }
-}
+// async function testTransaction() {
+//     const mongodb = global.mongodb
+//     const session = await global.getMongoSession()
+//     try {
+//         await mongodb.collection(Util.CollectionEnum._seq).findOneAndUpdate(
+//             { seqName: 'billSeq' },
+//             { $inc: { seqValue: 1 } },
+//             { returnOriginal: false, projection: { seqValue: 1, _id: 0 }, session }
+//         )
+//         await mongodb.collection(Util.CollectionEnum.bill).insertOne({ id: 1, sourceId: 1 }, { session })
+//         await mongodb.collection(Util.CollectionEnum.bill).insertOne({ id: 2 }, { session })
+//         await mongodb.collection(Util.CollectionEnum.bill).insertOne({ id: 3 }, { session })
+//         await mongodb.collection(Util.CollectionEnum.bill).insertOne({ id: 4, sourceId: 1 }, { session })
+//         await session.commitTransaction()
+//     } catch (error) {
+//         console.error('事务回滚')
+//         console.error(error.codeName)
+//         console.error(error.errmsg) // sourceId_1
+//         console.error('错误详情')
+//         console.error(error)
+//         await session.abortTransaction()
+//     } finally {
+//         await session.endSession()
+//     }
+// }
 
 module.exports = router
